Streamline audio data collection and processing on Android and iOS
- Updated C++ `TunerCPP` class to accept raw pointers (`const float*`, `const int16_t*`) to minimize array copying and type casting.
- Optimized Android JNI bridge to pass `ShortArray` directly to C++.
- Removed expensive `map` and `toDoubleArray` conversions in Android Kotlin code.
- Optimized iOS Swift code to use a `Float` buffer and pass data via `UnsafeBufferPointer` to the Objective-C++ bridge.
- Unified audio processing logic in C++ with normalization handled where necessary.
- Updated Dart tests to align with the current plugin interface.

+++ ios/Classes/TunerBridge.h
@@ -5,7 +5,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface TunerBridge : NSObject
 
 - (instancetype)initWithSampleRate:(int)sampleRate bufferSize:(int)bufferSize;
-- (double)findFrequency:(const double *)data length:(int)length;
+- (double)findFrequency:(const float *)data length:(int)length;
 - (void)dispose;
 
 @end
